+++ packages/accounts-base/accounts_server.js
@@ -364,6 +364,7 @@ Ap._attemptLogin = function (
       ),
       result.options || {}
     );
+    ret.type = attempt.type;
     this._successfulLogin(methodInvocation.connection, attempt);
     return ret;
   }