+++ packages/accounts-base/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   summary: "A user account system",
-  version: "1.4.1"
+  version: "1.4.2"
 });
 
 Package.onUse(function (api) {